Move RURuntimeStats out of internal pkg
